fix(ranking): cast count(*) results to int in raw queries
Prisma $queryRaw returns Postgres bigint counts as JS BigInt at runtime,
while the declared generic types claim number. Intl.PluralRules.select()
then throws "Conversion from 'BigInt' to 'number' is not allowed" when
formatting /ranking serwer entries. Cast count(*) and count(distinct ...)
to int4 so runtime values match the declared types.

+++ apps/bot/src/ranking.ts
@@ -378,8 +378,8 @@ export const ranking = new Hashira({ name: "ranking" })
                   >`
                   select
                     "channelId",
-                    count(*) as "total",
-                    count(distinct "userId") as "uniqueMembers"
+                    count(*)::int as "total",
+                    count(distinct "userId")::int as "uniqueMembers"
                   from "userTextActivity"
                   where
                     "guildId" = ${itx.guild.id}
@@ -473,8 +473,8 @@ export const ranking = new Hashira({ name: "ranking" })
                 >`
                   select
                     "userId",
-                    count(*) as "total",
-                    count(distinct "channelId") as "uniqueChannels"
+                    count(*)::int as "total",
+                    count(distinct "channelId")::int as "uniqueChannels"
                   from "userTextActivity"
                   where
                     "guildId" = ${itx.guild.id}
